pyproject.toml
[project]
name = "ai-operations-engine"
version = "0.1.0"
description = "AI-assisted operations engine with deterministic rules, operational memory and LLM reasoning."
requires-python = ">=3.10"
dependencies = [
    "pyyaml"
]